Model inference server and deployment/requirements.txt
flask==2.0.1
werkzeug==2.0.1
tensorflow==2.9.1
Pillow==8.3.1
numpy==1.21.0
h5py==3.6.0
requests==2.28.1
